Implement endpoint handler that controls HTTP interaction logging

+++ src/main/java/org/prebid/server/log/model/HttpLogSpec.java
@@ -11,7 +11,7 @@ public class HttpLogSpec {
 
     String account;
 
-    long limit;
+    int limit;
 
     public enum Endpoint {
         auction, amp
